requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.4.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-extensions-shell==1.7.4.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
executing==0.8.3
gunicorn==20.1.0
idna==3.3
ipython==8.2.0
jedi==0.18.1
Markdown==3.3.4
martor==1.6.13
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
python-decouple==3.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.1.0
